e2e tests infra for compatibility testing
